vulnb0x/main-app/requirements.txt
flask
python-dotenv
flask_socketio
gevent
gevent-websocket
dataclasses-json
pymongo
redis
bcrypt
dacite
GitPython
docker